+++ src/main/java/lotto/model/Lotto.java
@@ -1,8 +1,9 @@
-package lotto;
+package lotto.model;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import lotto.model.number.WinningNumbers;
 
 public class Lotto {
     private static final List<Integer> NOMINEE_NUMBERS = new ArrayList<>();
