+++ src/test/java/lotto/model/number/WinningNumbersTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model.number;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
