+++ src/test/java/lotto/model/prize/PrizeInformationsTest.java
@@ -1,8 +1,9 @@
-package lotto;
+package lotto.model.prize;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import lotto.model.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
